% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants_mean_fun.R
\name{quants_mean_fun}
\alias{quants_mean_fun}
\title{Function for means and CIs}
\usage{
quants_mean_fun(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
named list with quantiles \code{quants} and the mean \code{mean}
}
\description{
Function for means and CIs
}
